DjangoREST/requirements.txt
Django==2.2.6
django-filter==2.2.0
djangorestframework==3.10.3
Markdown==3.1.1
Pillow==6.2.0
pytz==2019.3
sqlparse==0.3.0
